requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
python-dotenv==1.1.1
Requests==2.32.5
#boto3==1.40.56
#botocore==1.40.56
